refactor + login fix
