+++ public/js/paintingBoard/projects.js
@@ -1,4 +1,4 @@
-import { asyncRequest, parseLayers } from '/js/functions.js';
+import { asyncRequest, parseLayers } from '/js/utils/functions.js';
 
 const projectBoard = document.getElementById('projectBoard');
