+++ views/partials/navbar.ejs
@@ -1,15 +1,13 @@
+<script type="module" src="/js/navbar.js"></script>
 <nav id="mainNavbar">
   <a class="mainNavbar-brand" href="#">JAES</a>
   <button class="navbar-button" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav"
     aria-expanded="false" aria-label="Toggle navigation">
     &#9776;
   </button>
-  <script type="module" src="/js/navbar.js"></script>
-    
-  </script>
   <ul class="navbarMenu collapsable" id="navbarMenu">
     <li class="nav-item">
-      <a class="nav-link" href="/">Home<!--span class="sr-only"> (current)</span--></a>
+      <a class="nav-link" href="/">Home</a>
     </li>
     <li class="nav-item">
       <a class="nav-link" href="/game">Game</a>
@@ -24,7 +22,6 @@
       <a class="nav-link" href="/game/admin">Admin</a>
     </li>
   </ul>
-  </div>
   <ul class="navbarMenu" id="navbarMenuRight">
     <li class="nav-item nologged">
       <a href="/login" class="nav-link">Login</a>
