+++ views/user/register.ejs
@@ -1,7 +1,6 @@
 <% layout('layout') %>
 
 <div class="form-register" id="form-register">
-    <img class="mb-4" src="/docs/4.5/assets/brand/bootstrap-solid.svg" alt="" width="72" height="72">
     <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
 
     <div class="form-group">
